Loop multiple Read if they do not fill all the requested buffer
With @Maxxen.

Issue that lead to look into this was that shapelib bails out if a read do not
fill the buffer completely, while GDALFileSystem (and other filesystems downstream)
will potentially return only partial reads (case in point that triggered that was the duckdb-wasm filesystem).
Solution is to make so that GDAL Read takes responsibility for filling the buffer.